Ensure that gRPC pipes via pointer files behave the same as TCP connections when there is no server available
The behaviour of a gRPC client when a server is not available is for the calls to fail with StatusCode.Unavailable. Previously we'd throw FileNotFoundException, but this causes dependency injection failures, where code would otherwise expect the gRPC client to be injected successfully, but then be handling connection failures when the calls are made.