Replace ManualResetEventSlim since it might be buggy according to the automation trace logs
The automation trace logs show one of the workers proceeding past the ReadyForTesting.Wait(0) even though there's no "Ready to begin testing for descriptor" entry, which is necessary for the ManualResetEventSlim to ever be set. Replace ManualResetEventSlim with our own implementation of a gate so that we know it behaves as we want it to.