Implement "Commandlet" plugin test type, allow dynamic providers to directly specify implementation
Previously dynamic config providers that wanted to run some UET behaviour as part of the BuildGraph step had to define a custom command inside the main UET project, and have BuildGraph call back into it via the 'internal' subcommand (e.g. 'run-automation-test-from-buildgraph').

This doesn't scale well as we add more providers for testing and deployment, as each provider had to manually serialize the relevant settings for the command-line, and then deserialize them when the custom 'internal' command ran.

Instead, providers can now implement IDynamicReentrantExecutor to specify the code to run as part of the BuildGraph node, and use WriteDynamicReentrantSpawnAsync to write the BuildGraph nodes necessary to call back into UET and get the IDynamicReentrantExecutor code running.