Don't run functional tests on the CI server for the moment
They rely on the published executables of UET, which aren't ready at the time the normal tests are running (and can't be built on the fly due to build files being written to at the same time).

We should probably run the functional tests overnight by having the CI server pull the latest published binaries and running the tests against those.