Remove Git mounting support from UEFS for now
I have a strong suspicion that libgit2sharp might be the cause of our AccessViolationExceptions on macOS. To check this, I'm removing all of the code dependent on libgit2sharp from UEFS, which includes the current Git mounting code.

At the moment, the Git mounting code isn't production ready anyway. It's way too slow for performant use, and we need to finish Redpoint.Git.Managed and use that instead before we add support back.