Remove use of StreamCopier from LocalPackageFs
This implementation looks suspicious in that it doesn't seem to handle exceptions inside PerformReads/PerformWrites. If an exception were to occur there, I think it's possible the pull operation might stall on macOS.

Just get rid of this ring-based copier for now. It was designed so that it could be reading while writing, but in practice I think we want to prefer simple, reliable code for this operation given that it only happens once per package per machine.