wip: Shelving UTF-8 lexer work
This would add UTF-8 APIs to Redpoint.Lexer so we can memory map the header files and lex them directly. However, the added complexity is unlikely to be worth the performance improvement, given that the UTF-8 -> UTF-16 conversion only needs to be done once when reading the file, and we will have to read the entirety of header files anyway to find all the relevant preprocessor directives.