Add 'uet android keep-wireless-enabled' command (#117)
* Add 'uet android keep-wireless-enabled' command

This command should be run on a machine that has Android devices connected via USB. When run, it will switch any devices connected over USB to listen on TCP/IP 5555 so they can be used by the whole build cluster over Wi-Fi. It also forces the public/private key pair for the local user profile to be a well-known key pair, so that USB connections always have the same RSA key for devices (thus avoiding re-authorization prompts).

* Set adbkey on all builds, not just during KWE command

It looks like running `adb connect ip:5555` from another machine still requires the `adbkey` to be set - i.e. the ADB key is not just used for USB debugging, but authorization over Wi-Fi as well. Just make all build and developer machines use the same key to avoid the authorization nonsense. We'd prefer to put our key at a unique path for UET and pass the key path via an environment variable or `adb connect`, but ADB doesn't support this so we just have to override the key globally.

* Only require Android SDK for KWE command