Fix inconsistency of plugin code build when packaging for Marketplace/Fab (#126)
This fixes an issue where the editor would not be built if there were no tests to run for a package plugin, even though the editor was explicitly included in the build list for the distribution in BuildConfig.json.

As of this change, the build steps are always run based on what's in BuildConfig.json, but if the package type is set to 'None', then the binaries from these build steps don't get included in the packaged plugin for tests (so they'll be built again).