Remove old --storage-virtualisation flag
The code paths that this flag turns on haven't been tested (or even working) in a long time. The intent of this flag was to use UEFS to perform "copy-on-write" snapshots of working directories, so that the local build executor could build things in parallel. However, the difficulty of getting copy-on-write mounts in UEFS with correct NTFS semantics proved too difficult, and Redpoint internally just has the local build executor run in sequence for local development scenarios (CI builds are still distributed via GitLab).