Add 'run-containerd' for running containerd as a separate service
This adds a separate 'uet cluster run-containerd' command which is responsible for running containerd. It retrieves the containerd manifest from a cache on disk, and connects to ws://127.0.0.1:8375/containerd. It is intended that the main RKM service will offer a WebSocket that pushes the manifest out whenever it changes - the main RKM service pulling this information from the cluster controller, which in turn should use Kubernetes as an authoritative version (with an on-disk cache).

The benefit of doing this is that we can allow cluster administrators to change the containerd version on all of the cluster nodes, without manually changing configuration on each machine.

In addition, this command is much more efficient at downloading and extracting containerd and it's dependencies, and starts much faster than the main RKM service currently does.

The main RKM service has not yet been updated to use the new 'run-containerd' command yet.